Split out internal trait logic into sealed module.
